Get rid of eslint-config-standard
The main reason I want to remove this dependency is its dependency tree
is in my opinion untenable and unacceptable.

As an example the configuration depends on eslint-plugin-import, which
we don't even us anywhere because we use a different (and better)
import-sorting plugin.

Despite us not using it the plugin has to be installed in order for
ESLint to run successfully (it's referenced in eslint-config-standard).

The plugin depends on a bunch packages, some of which are seemingly very
small but they have a lot of dependencies on their own, the total number
and size of which may surprise you.

Merely remove the eslint-plugin-import dependency from the Lune backend
project generates an 1794-line yarn.lock diff. The node_modules
directory shrinks by 147 MB (-20%) and the number of files in it drops by
almost 32 thousand (-37%) (roughly, measured with find node_modules | wc -l).

I think we should stay far away from all of this.

Lastly, I don't believe it gives us anything valuable on top of what we
already define directly or import from somewhere else. If it turns out
we're missing some of it we can add pieces as needed.