Disable the no-use-before-define rule
What it says in the comment, we don't find this rule particularly
valuable and it generates errors in the most inconvenient scenarios
where we have to move code around (bloats the diff sizes) or add
eslint-disable-next-line pragmas to the code, neither of which is nice.

Already disables in at least some of our projects, this patch attempts
to centralize that.